requirements.txt
Django==3.0.1
django-import-export==2.0   
djangorestframework==3.11.0
python-decouple==3.3
dj-database-url==0.5.0
dj-static==0.0.6
psycopg2-binary==2.8.4
gunicorn==20.0.4